Backport Cuckoo hash fix from GitHub, bringing Gitlab up to date with Github

See merge request marc/swanky!51

+++ scuttlebutt/benches/block.rs
@@ -14,7 +14,7 @@ use scuttlebutt::{AesRng, Block};
 fn bench_hash_pt(c: &mut Criterion) {
     c.bench_function("Block::hash_pt", |b| {
         let pt = RistrettoPoint::random(&mut rand::thread_rng());
-        let tweak = rand::random::<usize>();
+        let tweak = rand::random::<u128>();
         b.iter(|| {
             let h = Block::hash_pt(black_box(tweak), black_box(&pt));
             black_box(h)
